\name{TERGMundirStab-package}
\alias{TERGMundirStab-package}
\alias{TERGMundirStab}
\docType{package}
\title{
\packageTitle{TERGMundirStab}
}
\description{
\packageDescription{TERGMundirStab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TERGMundirStab}
\packageIndices{TERGMundirStab}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TERGMundirStab}

Maintainer: \packageMaintainer{TERGMundirStab}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
